+++ src/main/java/fun/mike/flapjack/pipeline/lab/OutputChannel.java
@@ -6,7 +6,7 @@ import java.util.List;
 import fun.mike.record.alpha.Record;
 
 public interface OutputChannel<T> extends AutoCloseable {
-    boolean receive(Long number, String line, Record value);
+    boolean receive(int number, String line, Record value);
 
     default List<PipelineError> getErrors() {
         return Collections.emptyList();
